+++ packages/terradart_google/lib/project.dart
@@ -2,4 +2,6 @@
 /// Project services (enabling individual GCP APIs on a project).
 library;
 
+export 'src/project/apis.dart' show Apis;
+export 'src/project/barrels.dart' show Barrels;
 export 'src/project/google_project_service.dart' show GoogleProjectService;